feat: DistributionMap full screen optimize
