+++ apps/web/src/modules/analyze/DataView/OverviewSummary/index.tsx
@@ -37,10 +37,45 @@ const Overview: React.FC<{
 };
 
 const OverviewSummary = () => {
-  const { items } = useMetricQueryData();
+  const { items, loading } = useMetricQueryData();
+  if (loading) {
+    return <Loading />;
+  }
   return <Overview data={items} />;
 };
-
+const Loading = () => (
+  <div className="h-[430px] animate-pulse rounded border bg-white p-10 px-6 py-6 shadow">
+    <div className="flex-1 space-y-4">
+      <div className="h-4 rounded bg-slate-200"></div>
+      <div className="grid grid-cols-3 gap-4">
+        <div className="col-span-2 h-4 rounded bg-slate-200"></div>
+        <div className="col-span-1 h-4 rounded bg-slate-200"></div>
+      </div>
+      <div className="h-4 rounded bg-slate-200"></div>
+      <div className="grid grid-cols-3 gap-4">
+        <div className="col-span-1 h-4 rounded bg-slate-200"></div>
+        <div className="col-span-2 h-4 rounded bg-slate-200"></div>
+      </div>
+      <div className="h-4 rounded bg-slate-200"></div>
+      <div className="h-4 rounded bg-slate-200"></div>
+      <div className="grid grid-cols-3 gap-4">
+        <div className="col-span-2 h-4 rounded bg-slate-200"></div>
+        <div className="col-span-1 h-4 rounded bg-slate-200"></div>
+      </div>
+      <div className="h-4 rounded bg-slate-200"></div>
+      <div className="grid grid-cols-3 gap-4">
+        <div className="col-span-1 h-4 rounded bg-slate-200"></div>
+        <div className="col-span-2 h-4 rounded bg-slate-200"></div>
+      </div>
+      <div className="h-4 rounded bg-slate-200"></div>
+      <div className="grid grid-cols-3 gap-4">
+        <div className="col-span-1 h-4 rounded bg-slate-200"></div>
+        <div className="col-span-2 h-4 rounded bg-slate-200"></div>
+      </div>
+      <div className="h-4 rounded bg-slate-200"></div>
+    </div>
+  </div>
+);
 export default withErrorBoundary(OverviewSummary, {
   FallbackComponent: ErrorFallback,
   onError(error, info) {
