+++ apps/web/src/modules/auth/components/Copyright.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 const Copyright = () => {
   return (
     <div className="fixed bottom-12 w-full px-6 text-center text-xs text-slate-600">
-      <p className="">Copyright©2023 OSS Compass. All Rights Reserved.</p>
+      <p className="">Copyright©2024 OSS Compass. All Rights Reserved.</p>
     </div>
   );
 };
